Move P2C into a submodule
