+++ tower-balance/examples/demo.rs
@@ -55,7 +55,7 @@ fn main() {
     let fut = future::lazy(move || {
         let decay = Duration::from_secs(10);
         let d = gen_disco();
-        let pe = lb::P2CBalance::from_entropy(load::PeakEwmaDiscover::new(
+        let pe = lb::p2c::Balance::from_entropy(load::PeakEwmaDiscover::new(
             d,
             DEFAULT_RTT,
             decay,
@@ -67,7 +67,7 @@ fn main() {
     let fut = fut.then(move |_| {
         let d = gen_disco();
         let ll =
-            lb::P2CBalance::from_entropy(load::PendingRequestsDiscover::new(d, load::NoInstrument));
+            lb::p2c::Balance::from_entropy(load::PendingRequestsDiscover::new(d, load::NoInstrument));
         run("P2C+LeastLoaded...", ll)
     });
 
@@ -129,7 +129,7 @@ fn gen_disco() -> impl Discover<
     )
 }
 
-fn run<D>(name: &'static str, lb: lb::P2CBalance<D>) -> impl Future<Item = (), Error = ()>
+fn run<D>(name: &'static str, lb: lb::p2c::Balance<D>) -> impl Future<Item = (), Error = ()>
 where
     D: Discover + Send + 'static,
     D::Error: Into<Error>,
